+++ tests/test_main.py
@@ -103,13 +103,13 @@ def _counts_to_inch(counts: int) -> float:
 
 
 def test_parse_response():  # noqa: PLR0915
-    units = wlll.units.Units()
+    units = wlll.Units()
     conditions = wlll.parse_response(RESPONSE, units)
 
     assert conditions.timestamp == datetime.fromtimestamp(1531754005, timezone.utc)
 
     inside = conditions.inside
-    assert isinstance(inside, wlll.conditions.InsideConditions)
+    assert isinstance(inside, wlll.InsideConditions)
     assert inside.lsid == 48307
     assert inside.temp == 78
     assert inside.hum == 41.1
@@ -117,7 +117,7 @@ def test_parse_response():  # noqa: PLR0915
     assert inside.heat_index == 8.4
 
     barometric = conditions.barometric
-    assert isinstance(barometric, wlll.conditions.BarometricConditions)
+    assert isinstance(barometric, wlll.BarometricConditions)
     assert barometric.lsid == 48306
     assert barometric.bar_absolute == 30.008
     assert barometric.bar_sea_level == 30.008
@@ -125,7 +125,7 @@ def test_parse_response():  # noqa: PLR0915
 
     assert len(conditions.moisture_temperature_stations) == 1
     mt0 = conditions.moisture_temperature_stations[0]
-    assert isinstance(mt0, wlll.conditions.MoistureTemperatureConditions)
+    assert isinstance(mt0, wlll.MoistureTemperatureConditions)
     assert mt0.txid == 3
     assert mt0.rx_state is None
     assert mt0.trans_battery_flag is None
@@ -142,10 +142,10 @@ def test_parse_response():  # noqa: PLR0915
 
     assert len(conditions.integrated_sensor_suites) == 1
     iss0 = conditions.integrated_sensor_suites[0]
-    assert isinstance(iss0, wlll.conditions.SensorSuiteConditions)
+    assert isinstance(iss0, wlll.SensorSuiteConditions)
     assert iss0.lsid == 48308
     assert iss0.txid == 1
-    assert iss0.rx_state == wlll.conditions.RadioReceptionState.SCANNING
+    assert iss0.rx_state == wlll.RadioReceptionState.SCANNING
     assert iss0.trans_battery_flag == 0
     assert iss0.temp == 62.7
     assert iss0.hum == 1.1
@@ -190,11 +190,11 @@ def test_get_conditions():
         body=RESPONSE,
     )
 
-    units = wlll.units.Units(
-        temperature=wlll.units.TemperatureUnit.CELSIUS,
-        pressure=wlll.units.PressureUnit.HECTOPASCAL,
-        rain=wlll.units.RainUnit.MILLIMETER,
-        wind_speed=wlll.units.WindSpeedUnit.METER_PER_SECOND,
+    units = wlll.Units(
+        temperature=wlll.TemperatureUnit.CELSIUS,
+        pressure=wlll.PressureUnit.HECTOPASCAL,
+        rain=wlll.RainUnit.MILLIMETER,
+        wind_speed=wlll.WindSpeedUnit.METER_PER_SECOND,
     )
     conditions = wlll.get_conditions("127.0.0.1", units)
     assert conditions == wlll.parse_response(RESPONSE, units)